requirements.txt
pytesseract
Pillow
click
opencv-python
